new branch / refactor modules
I've split my code into user and exercise folders and refactored
accordingly. This should fix some problems with required many different
files.
