+++ assets/scripts/exercise/events.js
@@ -1,0 +1,3 @@
+'use strict'
+const addExercise = require('./exerciseActions/addExercise')
+  $('#addExerciseButton').on('submit', addExercise.check)
