+++ assets/scripts/exercise/indexExercises.js
