--- assets/scripts/signIn.js
@@ -1,16 +1,0 @@
-'use strict'
-const ui = require('./ui')
-const config = require('./config')
-
-const onSignIn = (data) => {
-  return $.ajax({
-    url: config.apiOrigin + 'sign-in',
-    method: 'POST',
-    data
-  })
-  .done(ui.signInSuccess)
-  .fail(ui.signInFail)
-}
-module.exports = {
-  onSignIn
-}
