+++ assets/scripts/user/api.js
@@ -1,0 +1,48 @@
+'use strict'
+const config = require('./config')
+const store = require('./store')
+
+const onSignUp = (data) => {
+  return $.ajax({
+    url: config.apiOrigin + 'sign-up',
+    method: 'POST',
+    data
+  })
+}
+
+const onSignIn = (data) => {
+  return $.ajax({
+    url: config.apiOrigin + 'sign-in',
+    method: 'POST',
+    data
+  })
+}
+
+const onSignOut = () => {
+  return $.ajax({
+    url: config.apiOrigin + 'sign-out/' + store.user.id,
+    method: 'DELETE',
+    headers: {
+      Authorization: 'Token token=' + store.user.token
+    }
+  })
+}
+
+// changes password and stores credentials
+const onChangePassword = function (data) {
+  return $.ajax({
+    url: config.apiOrigin + 'change-password/' + store.user.id,
+    method: 'PATCH',
+    headers: {
+      Authorization: 'Token token=' + store.user.token
+    },
+    data
+  })
+}
+
+module.exports = {
+  onSignUp,
+  onSignIn,
+  onSignOut,
+  onChangePassword
+}
