+++ assets/scripts/user/showHide.js
