+++ index.html
@@ -142,12 +142,7 @@
   <!-- change password button -->
   <button type="button" id="changePasswordButton" class="btn btn-primary btn-lg modalb" data-toggle="modal" data-target="#change-pass-modal"> Change Password
               </button>
-  <!-- <div>
-    <p id="changePassNotification">
-    </p>
-    <p id="gameStatsNotification">
-    </p>
-  </div> -->
+
   <!-- sign out button -->
   <button type="button" id="sign-out" class="btn btn-primary btn-lg">
                 Sign Out
@@ -157,12 +152,10 @@
 
 <nav id="exerciseButtons">
   <!-- exercise data buttons -->
-  <button id="addExercisesButton" class="btn btn-primary btn-lg">
-    Add Exercise
-  </button>
-
+  <button type="button" id="addExerciseButton" class="btn btn-primary btn-lg modalb" data-toggle="modal" data-target="#addExerciseModal"> Add Exercise
+              </button>
 <!-- add exercise modal -->
-<div class="forms signUpHide">
+<div class="forms">
   <div class="row">
     <div class="col-md-12">
       <!-- Modal -->
